[ApportionmentFunction]: Fixed virtual method prototype
Should be double, not float.

+++ include/core/eval/ApportionmentFunction.hpp
@@ -14,7 +14,7 @@ class ApportionmentFunction : public EABehaviourObject {
 	public:
 	virtual ~ApportionmentFunction();
 
-	virtual float apportionFitness(
+	virtual double apportionFitness(
 		Genome* recipient,
 		Genome* provider,
 		unsigned int recipientStartIndex,
